changes size_t for int in ft_strlen && ft_strjoi

+++ get_next_line.h
@@ -7,7 +7,7 @@
 char	*get_next_line(int fd);
 char	*my_get_line(char *buffer, int start, int end);
 char	*reset_buffer(char *buffer);
-size_t	ft_strlen(const char *src);
+int	ft_strlen(const char *src);
 char	*ft_strjoin(char *s1, char *s2);
 int	ft_check_nl(char *line);
 char *ft_get_line(char * str);
